refactor(task): deduplicate TaskInput into shared task package
Move TaskInput to task package, alias in mcp/types.go, and remove
manual field-by-field mapping in handlers. Addresses Greptile review.
