+++ internal/task/models.go
@@ -23,6 +23,17 @@ const (
 	StatusReady      TaskStatus = "ready"       // Dependencies met, ready for execution
 )
 
+// TaskInput is a caller-provided task definition used to bypass LLM generation.
+// Shared between MCP handlers and the plan app layer.
+type TaskInput struct {
+	Title              string   `json:"title"`
+	Description        string   `json:"description,omitempty"`
+	AcceptanceCriteria []string `json:"acceptance_criteria,omitempty"`
+	ValidationSteps    []string `json:"validation_steps,omitempty"`
+	Priority           int      `json:"priority,omitempty"`
+	Complexity         string   `json:"complexity,omitempty"`
+}
+
 // PhaseStatus represents the lifecycle state of a phase
 type PhaseStatus string
